+++ src/com/pensio/api/request/ReserveSubscriptionChargeRequest.java
@@ -2,12 +2,29 @@ package com.pensio.api.request;
 
 import com.pensio.Amount;
 
-public class ReserveSubscriptionChargeRequest
+public class ReserveSubscriptionChargeRequest<T extends ReserveSubscriptionChargeRequest<T>>
 {
 	private String agreementId;
 	private AgreementUnscheduledType  agreementUnscheduledType;
 	private Amount amount;
-	
+    private PaymentInfos paymentInfos;
+
+    {
+        paymentInfos = new PaymentInfos();
+    }
+
+    public PaymentInfos getPaymentInfos()
+    {
+        return paymentInfos;
+    }
+
+    @SuppressWarnings("unchecked")
+    public T addPaymentInfo(String key, String value)
+    {
+        paymentInfos.add(key, value);
+        return (T)this;
+    }
+
 	public ReserveSubscriptionChargeRequest(String agreementId)
 	{
 		this.agreementId = agreementId;